+++ genomad/modules/annotate.py
@@ -55,7 +55,8 @@ def main(
     restart,
     threads,
     verbose,
-    conservative_taxonomy,
+    lenient_taxonomy,
+    full_ictv_lineage,
     sensitivity,
     evalue,
     splits,
@@ -229,7 +230,8 @@ def main(
             outputs.annotate_taxonomy_output,
             outputs.annotate_genes_output,
             database_obj,
-            conservative_taxonomy=conservative_taxonomy,
+            lenient_taxonomy=lenient_taxonomy,
+            full_ictv_lineage=full_ictv_lineage,
         )
         console.log(
             f"Taxonomic assignment data was written to [green]{outputs.annotate_taxonomy_output.name}[/green]."