+++ genomad/taxonomy.py
@@ -1,12 +1,44 @@
 from collections import defaultdict
 
 import taxopy
+
 from genomad import utils
 
 
+def get_conservative_taxon(taxon, taxdb):
+    for rank, taxid in taxon.ranked_taxid_lineage:
+        if rank not in {"subfamily", "genus", "subgenus", "species"}:
+            return taxopy.Taxon(taxid, taxdb)
+    return taxon
+
+
 def write_taxonomic_assignment(
-    taxonomy_output, genes_output, database_obj, conservative_taxonomy=False
+    taxonomy_output,
+    genes_output,
+    database_obj,
+    lenient_taxonomy=False,
+    full_ictv_lineage=False,
 ):
+    if full_ictv_lineage:
+        output_ranks = [
+            "realm",
+            "subrealm",
+            "kingdom",
+            "subkingdom",
+            "phylum",
+            "subphylum",
+            "class",
+            "subclass",
+            "order",
+            "suborder",
+            "family",
+        ]
+        if lenient_taxonomy:
+            output_ranks += ["subfamily", "genus", "subgenus", "species"]
+    else:
+        output_ranks = ["realm", "kingdom", "phylum", "class", "order", "family"]
+        if lenient_taxonomy:
+            output_ranks += ["genus", "species"]
     taxdb = database_obj.get_taxdb()
     contig_taxid_dict = defaultdict(lambda: ([], []))
     for line in utils.read_file(genes_output, skip_header=True):
@@ -25,19 +57,9 @@ def write_taxonomic_assignment(
                     taxon_list, taxdb, weights=bitscores, fraction=0.5
                 )
                 agreement = majority_taxon.agreement
-                # If classification is at the family level, be more conservative
-                if (
-                    majority_taxon.rank == "family"
-                    and agreement < 0.7
-                    and conservative_taxonomy
-                ):
-                    majority_taxon = taxopy.find_majority_vote(
-                        taxon_list, taxdb, weights=bitscores, fraction=0.7
-                    )
-                    agreement = majority_taxon.agreement
                 # If the contig was assigned to Nucleocytoviricota but contains
                 # at least one Caudoviricetes marker, be more conservative
-                elif (
+                if (
                     majority_taxon.rank_name_dictionary.get("phylum")
                     == "Nucleocytoviricota"
                     and agreement < 0.6
@@ -50,23 +72,34 @@ def write_taxonomic_assignment(
                         taxon_list, taxdb, weights=bitscores, fraction=0.6
                     )
                     agreement = majority_taxon.agreement
+                # If classification is below family level and `lenient_taxonomy`
+                # is False,
+                if not lenient_taxonomy and majority_taxon.rank in {
+                    "subfamily",
+                    "genus",
+                    "subgenus",
+                    "species",
+                }:
+                    majority_taxon = get_conservative_taxon(majority_taxon, taxdb)
+                    agreement = 0
+                    for t, w in zip(taxon_list, bitscores):
+                        if (
+                            t.rank_taxid_dictionary.get(majority_taxon.rank)
+                            == majority_taxon.taxid
+                        ):
+                            agreement += w / sum(bitscores)
             else:
                 majority_taxon = taxon_list[0]
                 agreement = 1
-                if majority_taxon.rank == "family" and conservative_taxonomy:
-                    majority_taxon = taxopy.Taxon(
-                        majority_taxon.taxid_lineage[1], taxdb
-                    )
+                if not lenient_taxonomy and majority_taxon.rank in {
+                    "subfamily",
+                    "genus",
+                    "subgenus",
+                    "species",
+                }:
+                    majority_taxon = get_conservative_taxon(majority_taxon, taxdb)
             lineage = [
-                majority_taxon.rank_name_dictionary.get(i, "")
-                for i in [
-                    "realm",
-                    "kingdom",
-                    "phylum",
-                    "class",
-                    "order",
-                    "family",
-                ]
+                majority_taxon.rank_name_dictionary.get(i, "") for i in output_ranks
             ]
             lineage = ";".join(["Viruses"] + lineage)
             fout.write(